backend2/requirements.txt
fastapi[standard]
pydantic
PyJWT
python-dotenv
psycopg2-binary
sqlmodel
sqlalchemy 
uvicorn
valkey
argon2-cffi
asgi_correlation_id
